Convert to colcon Cmake package and compile (#8)
