+++ amlip_node/VERSION
@@ -1,0 +1,3 @@
+VERSION_MAJOR 0
+VERSION_MINOR 0
+VERSION_PATCH 0
