+++ amlip_node/cmake/packaging/Config.cmake.in
@@ -1,0 +1,19 @@
+set(@PROJECT_NAME@_VERSION @PROJECT_VERSION@)
+
+@PACKAGE_INIT@
+
+if(MSVC OR MSVC_IDE)
+    set_and_check(@PROJECT_NAME@_BIN_DIR "@PACKAGE_BIN_INSTALL_DIR@")
+endif()
+set_and_check(@PROJECT_NAME@_INCLUDE_DIR "@PACKAGE_INCLUDE_INSTALL_DIR@")
+set_and_check(@PROJECT_NAME@_LIB_DIR "@PACKAGE_LIB_INSTALL_DIR@")
+
+if(NOT fastcdr_FOUND)
+    find_package(fastcdr)
+endif()
+
+if(NOT fastrtps_FOUND)
+    find_package(fastrtps)
+endif()
+
+include(${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@-targets.cmake)
