--- amlip_node/include/amlip_node/StatusAmlipNode.hpp
@@ -1,51 +1,0 @@
-// Copyright 2022 Proyectos y Sistemas de Mantenimiento SL (eProsima).
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/**
- * @file StatusAmlipNode.hpp
- */
-
-#ifndef AMLIP_AMLIPNODE_STATUSAMLIPNODE_HPP
-#define AMLIP_AMLIPNODE_STATUSAMLIPNODE_HPP
-
-#include <amlip_types/Status.hpp>
-
-namespace eprosima {
-namespace amlip {
-namespace node {
-
-class StatusAmlipNodeImpl;
-
-class StatusAmlipNode
-{
-public:
-
-    StatusAmlipNode();
-
-    virtual ~StatusAmlipNode();
-
-    void spin();
-
-    void stop();
-
-protected:
-
-    StatusAmlipNodeImpl* impl;
-};
-
-} /* namespace node */
-} /* namespace amlip */
-} /* namespace eprosima */
-
-#endif /* AMLIP_AMLIPNODE_STATUSAMLIPNODE_HPP */
