+++ amlip_node/include/library/config.h.in
@@ -1,0 +1,31 @@
+// Copyright 2022 Proyectos y Sistemas de Mantenimiento SL (eProsima).
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef _@SUBMODULE_PROJECT_MACROS@_LIBRARY_CONFIG_H_
+#define _@SUBMODULE_PROJECT_MACROS@_LIBRARY_CONFIG_H_
+
+#ifndef @SUBMODULE_PROJECT_MACROS@_VERSION_MAJOR
+#define @SUBMODULE_PROJECT_MACROS@_VERSION_MAJOR @PROJECT_VERSION_MAJOR@
+#endif // ifndef @SUBMODULE_PROJECT_MACROS@_VERSION_MAJOR
+#ifndef @SUBMODULE_PROJECT_MACROS@_VERSION_MINOR
+#define @SUBMODULE_PROJECT_MACROS@_VERSION_MINOR @PROJECT_VERSION_MINOR@
+#endif // ifndef @SUBMODULE_PROJECT_MACROS@_VERSION_MINOR
+#ifndef @SUBMODULE_PROJECT_MACROS@_VERSION_MICRO
+#define @SUBMODULE_PROJECT_MACROS@_VERSION_MICRO @PROJECT_VERSION_PATCH@
+#endif // ifndef @SUBMODULE_PROJECT_MACROS@_VERSION_MICRO
+#ifndef @SUBMODULE_PROJECT_MACROS@_VERSION_STR
+#define @SUBMODULE_PROJECT_MACROS@_VERSION_STR "@PROJECT_VERSION@"
+#endif // ifndef @SUBMODULE_PROJECT_MACROS@_VERSION_STR
+
+#endif // ifndef _@SUBMODULE_PROJECT_MACROS@_LIBRARY_CONFIG_H_
