+++ amlip_node/include/library/eProsima_auto_link.h.in
@@ -1,0 +1,71 @@
+// Copyright 2022 Proyectos y Sistemas de Mantenimiento SL (eProsima).
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+/*
+   Expected defines.
+   - EPROSIMA_@SUBMODULE_PROJECT_MACROS@_LIB_NAME
+   - @SUBMODULE_PROJECT_MACROS@_VERSION_MAJOR
+   - @SUBMODULE_PROJECT_MACROS@_VERSION_MINOR
+ */
+
+#ifndef _EPROSIMA_LIBRARY_EPROSIMAAUTOLINK_H_
+#define _EPROSIMA_LIBRARY_EPROSIMAAUTOLINK_H_
+
+#if defined(_MSC_VER)
+    #define EPROSIMA_STRINGIZE(X) EPROSIMA_DO_STRINGIZE(X)
+    #define EPROSIMA_DO_STRINGIZE(X) #X
+
+    #if defined(_DEBUG)
+        #define EPROSIMA_LIB_DEBUG_TAG "d"
+    #else
+        #define EPROSIMA_LIB_DEBUG_TAG
+    #endif // _DEBUG
+
+// Select linkage option.
+    #if (defined(_DLL) || defined(_RTLDLL)) && defined(EPROSIMA_DYN_LINK)
+        #define EPROSIMA_LIB_PREFIX
+    #elif defined(EPROSIMA_DYN_LINK)
+        #error "Mixing a dll eprosima library with a static runtime is a bad idea"
+    #else
+        #define EPROSIMA_LIB_PREFIX "lib"
+    #endif // if (defined(_DLL) || defined(_RTLDLL)) && defined(EPROSIMA_DYN_LINK)
+
+// Include library
+    #if defined(EPROSIMA_@SUBMODULE_PROJECT_MACROS@_LIB_NAME) \
+    && defined(EPROSIMA_LIB_PREFIX) \
+    && defined(EPROSIMA_LIB_DEBUG_TAG) \
+    && defined(@SUBMODULE_PROJECT_MACROS@_VERSION_MAJOR) \
+    && defined(@SUBMODULE_PROJECT_MACROS@_VERSION_MINOR)
+        #pragma \
+    comment(lib, EPROSIMA_LIB_PREFIX EPROSIMA_STRINGIZE(EPROSIMA_@SUBMODULE_PROJECT_MACROS@_LIB_NAME) EPROSIMA_LIB_DEBUG_TAG "-" EPROSIMA_STRINGIZE(@SUBMODULE_PROJECT_MACROS@_VERSION_MAJOR) "." EPROSIMA_STRINGIZE(@SUBMODULE_PROJECT_MACROS@_VERSION_MINOR) ".lib")
+    #else
+    #error "Some required macros where not defined"
+    #endif // if defined(EPROSIMA_@SUBMODULE_PROJECT_MACROS@_LIB_NAME) && defined(EPROSIMA_LIB_PREFIX) && defined(EPROSIMA_LIB_DEBUG_TAG) && defined(@SUBMODULE_PROJECT_MACROS@_VERSION_MAJOR) && defined(@SUBMODULE_PROJECT_MACROS@_VERSION_MINOR)
+
+#endif // _MSC_VER
+
+// Undef macros
+#ifdef EPROSIMA_LIB_PREFIX
+#undef EPROSIMA_LIB_PREFIX
+#endif // ifdef EPROSIMA_LIB_PREFIX
+
+#ifdef EPROSIMA_@SUBMODULE_PROJECT_MACROS@_LIB_NAME
+#undef EPROSIMA_@SUBMODULE_PROJECT_MACROS@_LIB_NAME
+#endif // ifdef EPROSIMA_@SUBMODULE_PROJECT_MACROS@_LIB_NAME
+
+#ifdef EPROSIMA_LIB_DEBUG_TAG
+#undef EPROSIMA_LIB_DEBUG_TAG
+#endif // ifdef EPROSIMA_LIB_DEBUG_TAG
+
+#endif // _EPROSIMA_LIBRARY_EPROSIMAAUTOLINK_H_
