+++ amlip_node/package.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0"?>
+<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
+<package format="3">
+  <name>amlip_node</name>
+  <version>0.0.0</version>
+  <description>
+    eprosima AML IP
+  </description>
+  <maintainer email="RaulSanchezMateos@eprosima.com">Raul Sánchez-Mateos</maintainer>
+  <maintainer email="javierparis@eprosima.com">Javier París</maintainer>
+  <maintainer email="juanlopez@eprosima.com">Juan López</maintainer>
+  <license file="LICENSE">Apache 2.0</license>
+
+  <url type="website">https://www.eprosima.com/</url>
+  <url type="bugtracker">https://github.com/eProsima/AML-IP/issues</url>
+  <url type="repository">https://github.com/eProsima/AML-IP</url>
+
+  <buildtool_depend>cmake</buildtool_depend>
+
+  <depend>fastrtps</depend>
+  <depend>ddsrouter_utils</depend>
+
+  <export>
+    <build_type>cmake</build_type>
+  </export>
+</package>
