+++ amlip_node/src/cpp/types/idl/amlip.idl
